+++ src/CommonLibrary/CommonLibrary/TimeUtils.h
@@ -1,11 +1,12 @@
 #ifndef _LIB_COMMON_TIME_UTILS_H_
 #define _LIB_COMMON_TIME_UTILS_H_
+
 namespace CommonLib
 {
 	namespace TimeUtils
 	{
-		long GetCurrentTimeMs();
-		long GetCurrentDate(long* pnTimeMs = 0);
+		uint32 GetCurrentTimeMs();
+		uint32 GetCurrentDate(uint32* pnTimeMs = 0);
 	}
 }
